Fixed #7 - values are now passed as unicode strings
Thanks to Dave McLain for the report and initial patch.

The previous behaviour was passing values like `0` as raw values, thus
preventing them to be displayed with
{% if value %} value="{{ value }}"{% endif %}. The new behaviour is what
django does.

Also fixed a minor issue with the select widget that used to overwrite
self.choices instead of chaining it with the choices passed as an argument.