Fixed #25 again -- Input widget turns attrs with value True to False
That means step and min in this test are turned into False, while they
should be 1. In Python integer/long/float 1 equals to True.

This breaks any Slider or anything built on top of RangeInput that uses
an attribute of 1. Attributes == True but from a non-boolean type are
converted to strings.

Thanks @maraujop for the heads up and the initial test.

Conflicts:

	floppyforms/tests/__init__.py